[FIX][IMP] project: add task # in portal view
As tasks are now used for a lot of use cases including tickets and support
it seems fair to include the task ID as reference in the portal view
for customers. It eases the communication process as people can directly
give the reference (aka ID).

This commit therefore adds a column in portal list view with the task ID
as reference. It also adds it in the portal form view, just after the
task name.

Done after feedback from internal migration.