[FIX] web_settings_dashboard: no redirect on social button click
Before this commit, when we click on social tags in the Settings,
it opens a pop-up while redirecting to the home page in the main window.
The redirection is not intended.

This is caused by href="#", which purpose is to show the pointer cursor when
hovering on the links. The issue is that href="#" redirects to the index page
on click, which is the home page.

This commit keeps the href while cancelling the redirection on click in the
main window.