[FIX] website_event_sale: pricelist show no discount
On the website when selling a ticket, a percentage discount may be
applied on the ticket price.

This percentage is calculated as:
  {event product price} - {event product price reduced} / {event product price}

So with a ticket price 20, a linked product priced 60 reduced to 30, we
would get a ticket price 10 on the website.

This commit hide the discounted price if the pricelist that is applied
was configured as "Discount included in the price".

opw-1807889
closes #23611